+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- work with tags on contact
+### Added
+- Tag::Collection for add / remove tag
+
+## [2.3.9] - 2020-07-07
 ### Security
 - Control access to Mautic::ConnectionsController
 ## [2.3.8] - 2020-06-24